Split out the bound box method based on branch.
